+++ yt/cpp/mapreduce/client/file_reader.cpp
@@ -5,6 +5,7 @@
 
 #include <yt/cpp/mapreduce/common/helpers.h>
 #include <yt/cpp/mapreduce/common/retry_lib.h>
+#include <yt/cpp/mapreduce/common/retry_request.h>
 #include <yt/cpp/mapreduce/common/wait_proxy.h>
 
 #include <yt/cpp/mapreduce/interface/config.h>
@@ -14,11 +15,6 @@
 #include <yt/cpp/mapreduce/interface/logging/yt_log.h>
 
 #include <yt/cpp/mapreduce/io/helpers.h>
-
-#include <yt/cpp/mapreduce/http/helpers.h>
-#include <yt/cpp/mapreduce/http/http.h>
-#include <yt/cpp/mapreduce/http/http_client.h>
-#include <yt/cpp/mapreduce/http/retry_request.h>
 
 namespace NYT {
 namespace NDetail {
