+++ yt/cpp/mapreduce/client/prepare_operation.cpp
@@ -1,8 +1,6 @@
 #include "prepare_operation.h"
 
-#include <yt/cpp/mapreduce/common/retry_lib.h>
-
-#include <yt/cpp/mapreduce/http/retry_request.h>
+#include <yt/cpp/mapreduce/common/retry_request.h>
 
 #include <yt/cpp/mapreduce/interface/raw_batch_request.h>
 #include <yt/cpp/mapreduce/interface/raw_client.h>
